requirements-dev.txt
asgiref==3.8.1
astroid==3.3.9
autopep8==2.3.2
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
crispy-bootstrap5==2024.10
diff-match-patch==20241021
dill==0.3.9
distlib==0.3.9
Django==5.1.5
django-admin-interface==0.29.4
django-appconf==1.1.0
django-colorfield==0.12.0
django-compressor==4.5.1
django-crispy-forms==2.3
django-filter==25.1
django-import-export==4.3.7
django-recaptcha==4.0.0
django-registration-redux==2.13
django-simple-history==3.8.0
django-tables2==2.7.5
django-tinymce==4.1.0
et_xmlfile==2.0.0
filelock==3.17.0
flake8==7.1.2
identify==2.6.9
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.3.5
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
razorpay==1.4.2
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
setuptools==76.0.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.3